+++ src/theory/ext_theory.cpp
@@ -67,23 +67,26 @@ std::ostream& operator<<(std::ostream& out, ExtReducedId id)
 }
 
 bool ExtTheoryCallback::getCurrentSubstitution(
-    int effort,
-    const std::vector<Node>& vars,
-    std::vector<Node>& subs,
-    std::map<Node, std::vector<Node> >& exp)
+    CVC5_UNUSED int effort,
+    CVC5_UNUSED const std::vector<Node>& vars,
+    CVC5_UNUSED std::vector<Node>& subs,
+    CVC5_UNUSED std::map<Node, std::vector<Node> >& exp)
 {
   return false;
 }
-bool ExtTheoryCallback::isExtfReduced(
-    int effort, Node n, Node on, std::vector<Node>& exp, ExtReducedId& id)
+bool ExtTheoryCallback::isExtfReduced(CVC5_UNUSED int effort,
+                                      Node n,
+                                      CVC5_UNUSED Node on,
+                                      CVC5_UNUSED std::vector<Node>& exp,
+                                      ExtReducedId& id)
 {
   id = ExtReducedId::SR_CONST;
   return n.isConst();
 }
-bool ExtTheoryCallback::getReduction(int effort,
-                                    Node n,
-                                    Node& nr,
-                                    bool& isSatDep)
+bool ExtTheoryCallback::getReduction(CVC5_UNUSED int effort,
+                                     CVC5_UNUSED Node n,
+                                     CVC5_UNUSED Node& nr,
+                                     CVC5_UNUSED bool& isSatDep)
 {
   return false;
 }