+++ src/theory/rep_set_iterator.h
@@ -206,11 +206,11 @@ class RepBoundExt
    * field we are considering, which indicates that
    * the iterator will terminate with a failure.
    */
-  virtual bool resetIndex(RepSetIterator* rsi,
-                          Node owner,
-                          size_t i,
-                          bool initial,
-                          std::vector<Node>& elements)
+  virtual bool resetIndex(CVC5_UNUSED RepSetIterator* rsi,
+                          CVC5_UNUSED Node owner,
+                          CVC5_UNUSED size_t i,
+                          CVC5_UNUSED bool initial,
+                          CVC5_UNUSED std::vector<Node>& elements)
   {
     return true;
   }
@@ -220,7 +220,10 @@ class RepBoundExt
    * with this bound has been given a complete interpretation
    * for type tn.
    */
-  virtual bool initializeRepresentativesForType(TypeNode tn) { return false; }
+  virtual bool initializeRepresentativesForType(CVC5_UNUSED TypeNode tn)
+  {
+    return false;
+  }
   /** get variable order
    * If this method returns true, then varOrder is the order
    * in which we want to consider variables for the iterator.
@@ -228,7 +231,8 @@ class RepBoundExt
    * and the RepSetIterator is free to choose a default
    * variable order.
    */
-  virtual bool getVariableOrder(Node owner, std::vector<size_t>& varOrder)
+  virtual bool getVariableOrder(CVC5_UNUSED Node owner,
+                                CVC5_UNUSED std::vector<size_t>& varOrder)
   {
     return false;
   }